requirements.txt
streamlit==1.44.1
duckdb==1.2.2
dlt==1.9.0
dbt-core==1.9.4
dbt-duckdb==1.9.2
python-dotenv==1.1.0
lxml==5.3.2
matplotlib==3.10.1